requirements.txt
torch
torchvision
numpy
scikit-learn
umap-learn
matplotlib
pyyaml
